+++ src/components/Header/index.js
@@ -21,7 +21,7 @@ const Header = () => (
         <Link to="/" className={header__link}>
           Home
         </Link>
-        <Link to="/app/profile" className={header__link}>
+        <Link to="/profile" className={header__link}>
           Profile
         </Link>
         <Link to="/blog" className={header__link}>
