--- src/components/Profile.js
@@ -1,19 +1,0 @@
-import React from "react"
-import View from "./View"
-import { getCurrentUser } from "../utils/auth"
-
-const Profile = () => {
-  const { name } = getCurrentUser()
-
-  return (
-    <View title="Your Profile">
-      <p>Welcome back to your profile, {name}!</p>
-      <p>
-        This is a client-only route. You could set up a form to save information
-        about a user here.
-      </p>
-    </View>
-  )
-}
-
-export default Profile
