+++ src/pages/app.js
@@ -1,7 +1,6 @@
 import React from "react"
 import { Router } from "@reach/router"
 import Layout from "../components/Layout"
-import Profile from "../components/Profile"
 import AppML from "../components/AppML"
 import Login from "../components/Login"
 import PrivateRoute from "../components/PrivateRoute"
@@ -12,7 +11,6 @@ const App = () => (
     <Status />
     <Router>
       <PrivateRoute path="/app/ml" component={AppML} />
-      <Profile path="/app/profile" />
       <Login path="/app/login" />
     </Router>
   </Layout>
